Little bit more condensing, actually hooking up addVideo endpoint
